Create README.md
Usage
To add an item to your day planner, enter the item in the text box and click the "Add" button.
To delete selected items, select the items in the list and click the "Delete" button.
To submit items, click the "Submit" button.
Customization
You can customize the appearance of the GUI and the font styles by modifying the code in day_planner.py. Feel free to make changes to suit your preferences.

Contributing
If you'd like to contribute to this project or report issues, please follow these guidelines:

Fork the repository on GitHub.
Make your changes and create a pull request.
License
This project is licensed under the MIT License - see the LICENSE file for details.

Acknowledgments
Special thanks to the tkinter library for making GUI development in Python easy.